requirements.txt
discord.py==1.3.3
beautifulsoup4==4.9.1